requirements.txt
# Web 框架
fastapi==0.115.0
uvicorn[standard]==0.32.0

# 数据库
supabase==2.10.0
psycopg2-binary==2.9.10
neo4j==5.27.0

# LangChain & LangGraph
langchain==0.3.13
langchain-core==0.3.28
langchain-community==0.3.13
langgraph==0.2.60
langchain-checkpoint-postgres>=2.0.0
langchain-openai==0.2.14

# 向量库
chromadb==0.5.23
sentence-transformers==3.3.1

# 网络搜索
duckduckgo-search

# Gradio
gradio==6.5.1

# 工具
python-dotenv==1.0.1
pydantic==2.10.4

# 文档处理
pypdf==5.1.0
unstructured==0.18.31

# 其他
tiktoken>=0.5.0
